% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albumin.R
\name{albumin}
\alias{albumin}
\title{Serum Albumin}
\usage{
albumin(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum albumin records
}
\description{
This function extracts serum albumin records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

Medcode  Readcode  Description
23       44M4.00   Serum albumin
13708    44MI.00   Plasma albumin level
26892    44M4000   Serum albumin normal
26893    44M4100   Serum albumin low
}
