% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilirubin.R
\name{bilirubin}
\alias{bilirubin}
\title{Serum Bilirubin}
\usage{
bilirubin(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum bilirubin records
}
\description{
This function extracts serum bilirubin records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

Medcode  Readcode  Description
59 \tab  44E..00 \tab   Serum bilirubin level
66 \tab      44EC.00 \tab   Serum total bilirubin level
1183 \tab    44E2.00 \tab   Serum bilirubin raised
13724 \tab  44E3.00 \tab   Total bilirubin
13725 \tab  44E9.00 \tab   Plasma total bilirubin level
26898 \tab  44E1.00 \tab   Serum bilirubin normal
37205 \tab    44E6.00 \tab   Serum bilirubin borderline
44363 \tab    44EZ.00 \tab   Serum bilirubin NOS
101070 \tab   4QCE.00 \tab   Bilirubin level
101501 \tab   4QCE000 \tab   Bilirubin direct measurement
}
