% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.R
\name{bp}
\alias{bp}
\title{Blood Pressure}
\usage{
bp(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all systolic and diastolic blood pressure records
}
\description{
This function extracts systolic and diastolic test results from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
