% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatine.R
\name{creatine}
\alias{creatine}
\title{Serum Creatine}
\usage{
creatine(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum creatine records
}
\description{
This function extracts serum creatine records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

Medcode  Readcode  Description
5        44J3.00   Serum creatine
3927     44J3300   Serum creatine raised
13736    44JF.00   Plasma creatinine level
26903    44J3200   Serum creatinine normal
27095    4Q40.00   Creatinine level
35545    44J3100   Serum creatinine low
42345    44J3z00   Serum creatinine NOS
45096    44JD.00   Corrected serum creatinine level
62062    44JC.00   Corrected plasma creatinine level
}
