% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfr.R
\name{gfr}
\alias{gfr}
\title{Glomerular Filtration Rate (GFR)}
\usage{
gfr(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all glomerular filtration rate records
}
\description{
This function extracts glomerular filtration rate test results from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
