% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggt.R
\name{ggt}
\alias{ggt}
\title{This function extracts gamma glutamyl transferase records from CPRD GOLD.
It assumes that your cohort contains a column named patid.}
\usage{
ggt(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all gamma glutamyl transferase records
}
\description{
Codes:
}
\details{
Medcode  Readcode  Description
5214     44G4.00   Gamma - G.T. level
5891     44G4100   Gamma glutamyl transferase level abnormal
13752    44G4000   Gamma glutamyl transferase level normal
13753    44G9.00   Serum gamma-glutamyl transferase level
13754    44G7.00   Plasma gamma-glutamyl transferase level
}
