% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.R
\name{height}
\alias{height}
\title{Height}
\usage{
height(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all height records
}
\description{
This function extracts height records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
