% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight.R
\name{weight}
\alias{weight}
\title{Weight}
\usage{
weight(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all weight records
}
\description{
This function extracts weight records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
