% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alanine.R
\name{alanine}
\alias{alanine}
\title{Alanine Aminotransferase (ALT)}
\usage{
alanine(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all alanine aminotransferase records
}
\description{
This function extracts alanine aminotransferase test results from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
