% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albumin.R
\name{albumin}
\alias{albumin}
\title{Serum Albumin}
\usage{
albumin(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum albumin records
}
\description{
This function extracts serum albumin records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

\tabular{lllll}{
Medcode \tab  \tab Readcode \tab  \tab Description\cr
23 \tab  \tab 44M4.00 \tab  \tab Serum albumin\cr
13708 \tab  \tab 44MI.00 \tab  \tab Plasma albumin level\cr
26892 \tab  \tab 44M4000 \tab  \tab Serum albumin normal\cr
26893 \tab  \tab 44M4100 \tab  \tab Serum albumin low\cr
}
}
