% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aspartate.R
\name{aspartate}
\alias{aspartate}
\title{Aspartate Aminotransferase (AST)}
\usage{
aspartate(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all aspartate aminotransferase records
}
\description{
This function extracts aspartate aminotransferase test results from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
