% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilirubin.R
\name{bilirubin}
\alias{bilirubin}
\title{Serum Bilirubin}
\usage{
bilirubin(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum bilirubin records
}
\description{
This function extracts serum bilirubin records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

\tabular{lllll}{
Medcode \tab  \tab Readcode \tab  \tab Description\cr
59 \tab  \tab 44E..00 \tab  \tab Serum bilirubin level\cr
66 \tab  \tab 44EC.00 \tab  \tab Serum total bilirubin level\cr
1183 \tab  \tab 44E2.00 \tab  \tab Serum bilirubin raised\cr
13724 \tab  \tab 44E3.00 \tab  \tab Total bilirubin\cr
13725 \tab  \tab 44E9.00 \tab  \tab Plasma total bilirubin level\cr
26898 \tab  \tab 44E1.00 \tab  \tab Serum bilirubin normal\cr
37205 \tab  \tab 44E6.00 \tab  \tab Serum bilirubin borderline\cr
44363 \tab  \tab 44EZ.00 \tab  \tab Serum bilirubin NOS\cr
101070 \tab  \tab 4QCE.00 \tab  \tab Bilirubin level\cr
101501 \tab  \tab 4QCE000 \tab  \tab Bilirubin direct measurement\cr
}
}
