% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilirubin.R
\name{bilirubin}
\alias{bilirubin}
\title{Serum Bilirubin}
\usage{
bilirubin(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum bilirubin records
}
\description{
This function extracts serum bilirubin records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

Medcode  Readcode  Description
59       44E..00   Serum bilirubin level
66       44EC.00   Serum total bilirubin level
1183     44E2.00   Serum bilirubin raised
13724    44E3.00   Total bilirubin
13725    44E9.00   Plasma total bilirubin level
26898    44E1.00   Serum bilirubin normal
37205    44E6.00   Serum bilirubin borderline
44363    44EZ.00   Serum bilirubin NOS
101070   4QCE.00   Bilirubin level
101501   4QCE000   Bilirubin direct measurement
}
