% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi.R
\name{bmi}
\alias{bmi}
\title{Body Mass Index}
\usage{
bmi(cohort, height, weight)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}

\item{height}{A dataframe extracted from the height() function}

\item{weight}{A dataframe extracted from the weight() function}
}
\value{
A dataframe containing all bmi records
}
\description{
This function calculates BMI records using height and weight results from
CPRD GOLD.
}
\details{
It assumes that your cohort contains a column named patid and that the weight
and height tables have already been extracted.
}
