% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creatine.R
\name{creatine}
\alias{creatine}
\title{Serum Creatine}
\usage{
creatine(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all serum creatine records
}
\description{
This function extracts serum creatine records from CPRD GOLD.
It assumes that your cohort contains a column named patid.
}
\details{
Codes:

\tabular{lllll}{
Medcode \tab  \tab Readcode \tab  \tab Description\cr
5 \tab  \tab 44J3.00 \tab  \tab Serum creatine\cr
3927 \tab  \tab 44J3300 \tab  \tab Serum creatine raised\cr
13736 \tab  \tab 44JF.00 \tab  \tab Plasma creatinine level\cr
26903 \tab  \tab 44J3200 \tab  \tab Serum creatinine normal\cr
27095 \tab  \tab 4Q40.00 \tab  \tab Creatinine level\cr
35545 \tab  \tab 44J3100 \tab  \tab Serum creatinine low\cr
42345 \tab  \tab 44J3z00 \tab  \tab Serum creatinine NOS\cr
45096 \tab  \tab 44JD.00 \tab  \tab Corrected serum creatinine level\cr
62062 \tab  \tab 44JC.00 \tab  \tab Corrected plasma creatinine level\cr
}
}
