% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggt.R
\name{ggt}
\alias{ggt}
\title{This function extracts gamma glutamyl transferase records from CPRD GOLD.
It assumes that your cohort contains a column named patid.}
\usage{
ggt(cohort)
}
\arguments{
\item{cohort}{A dataframe containing at least one column called patid}
}
\value{
A dataframe containing all gamma glutamyl transferase records
}
\description{
Codes:
}
\details{
\tabular{lllll}{
Medcode \tab  \tab Readcode \tab  \tab Description\cr
5214 \tab  \tab 44G4.00 \tab  \tab Gamma - G.T. level\cr
5891 \tab  \tab 44G4100 \tab  \tab Gamma glutamyl transferase level abnormal\cr
13752 \tab  \tab 44G4000 \tab  \tab Gamma glutamyl transferase level normal\cr
13753 \tab  \tab 44G9.00 \tab  \tab Serum gamma-glutamyl transferase level\cr
13754 \tab  \tab 44G7.00 \tab  \tab Plasma gamma-glutamyl transferase level\cr
}
}
